Update patch set 1
Patch Set 1: Code-Review-1

Dear Moshe, From testing point of view, it would make more sense to use the ONAP CLI than the direct rest calls. and ONAP CLI already delivers the CLI for performing most of the ONAP actions. so I would like you to consider the ONAP CLI for making ONAP related test cases. Here i observed that , what ONAP CLI does already, you are repeating here the same and it would be double effor if we continue this path.

Also as we are coming up with new architecture, i would suggest to make use of it instead of placing under vnftest.

in vnftest, its better to concentrate those test cases which not bound to ONAP.

Patch-set: 1