Update patch set 1
Patch Set 1:

> I don't understand your response to Kanagaraj's comments.  What is
 > your plan to incorporate the CLI? Are you updating this commit?

currently there are 3 test cases that call ONAP APIs with a REST client.
Kanagaraj suggests that we use OCLI instead of REST client and I agree. However, integrating the OCLI is a task ahead of us and will require some investigation. After it is integrated we can refactor the existing tests.
Note that only the call method will be changed, the parameter handling within the test cases as well as parsing of JSON responses are still different between Amsterdam and Beijing, therefore, versioned test cases are still needed - thus this code should be commited.

Patch-set: 1